@@ -2,7 +2,7 @@
   "name": "sr-meeseeks",
   "author": "Joel Rodrigues Moreira",
   "repository": "https://github.com/joeldatabox/mr-meeseeks",
-  "version": "0.11.8",
+  "version": "0.12.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -17,54 +17,55 @@
     "e2e": "ng e2e"
   },
   "peerDependencies": {
-    "@angular/animations": "^8.0.3",
-    "@angular/cdk": "^8.0.1",
-    "@angular/common": "^8.0.3",
-    "@angular/compiler": "^8.0.3",
-    "@angular/core": "^8.0.3",
-    "@angular/forms": "^8.0.3",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^8.0.1",
-    "@angular/platform-browser": "^8.0.3",
-    "@angular/platform-browser-dynamic": "^8.0.3",
-    "@angular/router": "^8.0.3",
-    "@covalent/core": "^2.0.2",
-    "angular2-text-mask": "^9.0.0",
-    "text-mask-addons": "^3.8.0",
-    "reflect-metadata": "^0.1.13",
-    "class-transformer": "^0.2.3",
-    "chart.js": "^2.8.0",
-    "ng2-logger": "^2.0.48",
-    "ngx-device-detector": "1.3.7",
-    "core-js": "2.5.4",
-    "rxjs": "^6.5.2",
-    "zone.js": "^0.9.1",
+    "@angular/animations": "8.2.4",
+    "@angular/cdk": "8.1.4",
+    "@angular/common": "8.2.4",
+    "@angular/compiler": "8.2.4",
+    "@angular/core": "8.2.4",
+    "@angular/forms": "8.2.4",
+    "@angular/http": "7.2.15",
+    "@angular/material": "8.1.4",
+    "@angular/platform-browser": "8.2.4",
+    "@angular/platform-browser-dynamic": "8.2.4",
+    "@angular/router": "8.2.4",
+    "@covalent/core": "2.1.1",
+    "angular2-text-mask": "9.0.0",
+    "text-mask-addons": "3.8.0",
+    "reflect-metadata": "0.1.13",
+    "class-transformer": "0.2.3",
+    "chart.js": "2.8.0",
+    "ng2-logger": "2.0.48",
+    "ngx-device-detector": "1.3.11",
+    "core-js": "3.2.1",
+    "rxjs": "6.5.2",
+    "zone.js": "0.10.2",
     "moment": "2.24.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.6",
-    "@angular/cli": "8.0.6",
-    "@angular/compiler-cli": "^8.0.3",
-    "@angular/language-service": "^8.0.3",
-    "@types/jasmine": "~3.3.13",
+    "@angular-devkit/build-angular": "~0.803.1",
+    "@angular/cli": "8.3.1",
+    "@angular/compiler-cli": "8.2.4",
+    "@angular/language-service": "8.2.4",
+    "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~12.0.10",
+    "@types/node": "~12.7.2",
     "codelyzer": "~5.1.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-coverage-istanbul-reporter": "2.1.0",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "moment-es6": "^1.0.0",
-    "ng-packagr": "^5.3.0",
-    "npm-install-peers": "^1.2.1",
+    "karma-jasmine-html-reporter": "1.4.2",
+    "moment-es6": "1.0.0",
+    "ng-packagr": "5.5.0",
+    "npm-install-peers": "1.2.1",
     "protractor": "~5.4.2",
     "ts-node": "~8.3.0",
-    "tsickle": "^0.35.0",
-    "tslint": "~5.18.0",
-    "typescript": "3.4.5"
+    "tsickle": "0.37.0",
+    "tslint": "~5.19.0",
+    "typescript": "3.5.3",
+    "npm-check-updates": "3.1.21"
   }
 }
