@@ -10,13 +10,19 @@ export class SrViewInjectorService {
   /**
    * Adds any component to the container
    * @param viewContainer -> instance of the ViewContainerRef
-   * @param component -> class of its component to be injected
+   * @param component -> class of its component to be injectedr
    */
   addFormIn(viewContainer: ViewContainerRef, component: any): ComponentRef<any> {
+    viewContainer.clear();
     const factory = this.factoryResolver.resolveComponentFactory(component);
+    const componentCreated = viewContainer.createComponent(factory);
+    return componentCreated;
+
+
+    /*const factory = this.factoryResolver.resolveComponentFactory(component);
     const componentCreated = factory.create(viewContainer.parentInjector);
     viewContainer.insert(componentCreated.hostView);
-    return componentCreated;
+    return componentCreated;*/
   }
 
   /**
